Better output messages

+++ scripts/patch_c.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
-import glob, comby as cby
-
-srcs = glob.glob('./for-wasp/**/*.c', recursive=True)
-srcs = srcs + glob.glob('./for-wasp/**/*.h', recursive=True)
+import os
+import sys
+import glob
+import comby
 
 patterns = [
         ('#include <gillian-c/gillian-c.h>', '#include "mockups.h"'),
@@ -23,16 +23,35 @@ patterns = [
         ('ASSUME (:[assertion]);','assume(:[assertion]);')
 ]
 
-# main
-comby = cby.Comby()
-for src in srcs:
-    print(f'Transforming {src}...')
+def main(argv=None):
+    if argv is None:
+        argv = sys.argv[1:]
 
-    with open(src, 'r') as f:
-        data = f.read()
+    src_dir = argv[0]
+    if not os.path.exists(src_dir):
+        print('Directory \'{src_dir}\' does not exist!')
+        return -1
 
-    for p in patterns:
-        data = comby.rewrite(data, p[0], p[1])
+    cby = comby.Comby()
 
-    with open(src, 'w') as f:
-        f.write(data)
+    sources = os.path.join(src_dir, '**/*.c')
+    headers = os.path.join(src_dir, '**/*.h')
+
+    srcs = glob.glob(sources, recursive=True) + \
+           glob.glob(headers, recursive=True)
+    for src in srcs:
+        print(f'Transforming {src}...')
+
+        with open(src, 'r') as f:
+            data = f.read()
+
+        for p in patterns:
+            data = cby.rewrite(data, p[0], p[1])
+
+        with open(src, 'w') as f:
+            f.write(data)
+
+    return 0
+
+if __name__ == '__main__':
+    sys.exit(main())
